Remove unnecessary imports and paths from urls.py
