+++ components/paddock/paddock/settings.py
@@ -88,7 +88,6 @@ INSTALLED_APPS = [
     "django_prometheus",
     "telemetry.apps.TelemetryConfig",
     "b4mad_racing_website.apps.B4madRacingWebsiteConfig",
-    "frontpage.apps.FrontpageConfig",
     "api.apps.ApiConfig",
     "markdown_deux",
 ]
